+++ vector/src/main/java/im/vector/app/features/createdirect/CreateDirectRoomActivity.kt
@@ -45,23 +45,23 @@ import im.vector.app.core.utils.allGranted
 import im.vector.app.core.utils.checkPermissions
 import im.vector.app.features.contactsbook.ContactsBookFragment
 import im.vector.app.features.contactsbook.ContactsBookViewModel
-import im.vector.app.features.userdirectory.KnownUsersFragment
-import im.vector.app.features.userdirectory.KnownUsersFragmentArgs
-import im.vector.app.features.userdirectory.UserDirectoryFragment
-import im.vector.app.features.userdirectory.UserDirectorySharedAction
-import im.vector.app.features.userdirectory.UserDirectorySharedActionViewModel
-import im.vector.app.features.userdirectory.UserDirectoryViewModel
+import im.vector.app.features.userdirectory.UserListFragment
+import im.vector.app.features.userdirectory.UserListFragmentArgs
+import im.vector.app.features.userdirectory.UserListSharedAction
+import im.vector.app.features.userdirectory.UserListSharedActionViewModel
+import im.vector.app.features.userdirectory.UserListViewModel
+import im.vector.app.features.userdirectory.UserListViewState
 import kotlinx.android.synthetic.main.activity.*
 import org.matrix.android.sdk.api.failure.Failure
 import org.matrix.android.sdk.api.session.room.failure.CreateRoomFailure
 import java.net.HttpURLConnection
 import javax.inject.Inject
 
-class CreateDirectRoomActivity : SimpleFragmentActivity() {
+class CreateDirectRoomActivity : SimpleFragmentActivity(), UserListViewModel.Factory {
 
     private val viewModel: CreateDirectRoomViewModel by viewModel()
-    private lateinit var sharedActionViewModel: UserDirectorySharedActionViewModel
-    @Inject lateinit var userDirectoryViewModelFactory: UserDirectoryViewModel.Factory
+    private lateinit var sharedActionViewModel: UserListSharedActionViewModel
+    @Inject lateinit var userListViewModelFactory: UserListViewModel.Factory
     @Inject lateinit var createDirectRoomViewModelFactory: CreateDirectRoomViewModel.Factory
     @Inject lateinit var contactsBookViewModelFactory: ContactsBookViewModel.Factory
     @Inject lateinit var errorFormatter: ErrorFormatter
@@ -71,37 +71,36 @@ class CreateDirectRoomActivity : SimpleFragmentActivity() {
         injector.inject(this)
     }
 
+    override fun create(initialState: UserListViewState, args: UserListFragmentArgs): UserListViewModel {
+        return userListViewModelFactory.create(initialState, args)
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         toolbar.visibility = View.GONE
-        sharedActionViewModel = viewModelProvider.get(UserDirectorySharedActionViewModel::class.java)
-        if (intent?.getBooleanExtra(BY_QR_CODE, false)!!) {
-            if (isFirstCreation()) { openAddByQrCode() }
-        } else {
-            sharedActionViewModel
-                    .observe()
-                    .subscribe { sharedAction ->
-                        when (sharedAction) {
-                            UserDirectorySharedAction.OpenUsersDirectory    ->
-                                addFragmentToBackstack(R.id.container, UserDirectoryFragment::class.java)
-                            UserDirectorySharedAction.Close                 -> finish()
-                            UserDirectorySharedAction.GoBack                -> onBackPressed()
-                            is UserDirectorySharedAction.OnMenuItemSelected -> onMenuItemSelected(sharedAction)
-                            UserDirectorySharedAction.OpenPhoneBook         -> openPhoneBook()
-                        }.exhaustive
-                    }
-                    .disposeOnDestroy()
-            if (isFirstCreation()) {
-                addFragment(
-                        R.id.container,
-                        KnownUsersFragment::class.java,
-                        KnownUsersFragmentArgs(
-                                title = getString(R.string.fab_menu_create_chat),
-                                menuResId = R.menu.vector_create_direct_room,
-                                isCreatingRoom = true
-                        )
-                )
-            }
+
+        sharedActionViewModel = viewModelProvider.get(UserListSharedActionViewModel::class.java)
+        sharedActionViewModel
+                .observe()
+                .subscribe { action ->
+                    when (action) {
+                        UserListSharedAction.Close  -> finish()
+                        UserListSharedAction.GoBack -> onBackPressed()
+                        is UserListSharedAction.OnMenuItemSelected -> onMenuItemSelected(action)
+                        UserListSharedAction.OpenPhoneBook -> openPhoneBook()
+                        UserListSharedAction.AddByQrCode -> openAddByQrCode()
+                    }.exhaustive
+                }
+                .disposeOnDestroy()
+        if (isFirstCreation()) {
+            addFragment(
+                    R.id.container,
+                    UserListFragment::class.java,
+                    UserListFragmentArgs(
+                            title = getString(R.string.fab_menu_create_chat),
+                            menuResId = R.menu.vector_create_direct_room
+                    )
+            )
         }
         viewModel.selectSubscribe(this, CreateDirectRoomViewState::createAndInviteState) {
             renderCreateAndInviteState(it)
@@ -129,22 +128,22 @@ class CreateDirectRoomActivity : SimpleFragmentActivity() {
         if (allGranted(grantResults)) {
             if (requestCode == PERMISSION_REQUEST_CODE_READ_CONTACTS) {
                 doOnPostResume { addFragmentToBackstack(R.id.container, ContactsBookFragment::class.java) }
-            } else if (requestCode == PERMISSION_REQUEST_CODE_LAUNCH_CAMERA && intent?.getBooleanExtra(BY_QR_CODE, false)!!) {
+            } else if (requestCode == PERMISSION_REQUEST_CODE_LAUNCH_CAMERA) {
                 addFragment(R.id.container, CreateDirectRoomByQrCodeFragment::class.java)
             }
         } else {
             Toast.makeText(baseContext, R.string.missing_permissions_error, Toast.LENGTH_SHORT).show()
-            if (requestCode == PERMISSION_REQUEST_CODE_LAUNCH_CAMERA && intent?.getBooleanExtra(BY_QR_CODE, false)!!) {
+            if (requestCode == PERMISSION_REQUEST_CODE_LAUNCH_CAMERA) {
                 finish()
             }
         }
     }
 
-    private fun onMenuItemSelected(action: UserDirectorySharedAction.OnMenuItemSelected) {
+    private fun onMenuItemSelected(action: UserListSharedAction.OnMenuItemSelected) {
         if (action.itemId == R.id.action_create_direct_room) {
             viewModel.handle(CreateDirectRoomAction.CreateRoomAndInviteSelectedUsers(
                     action.invitees,
-                    action.existingDmRoomId
+                    null
             ))
         }
     }
@@ -198,12 +197,9 @@ class CreateDirectRoomActivity : SimpleFragmentActivity() {
     }
 
     companion object {
-        private const val BY_QR_CODE = "BY_QR_CODE"
 
-        fun getIntent(context: Context, byQrCode: Boolean = false): Intent {
-            return Intent(context, CreateDirectRoomActivity::class.java).apply {
-                putExtra(BY_QR_CODE, byQrCode)
-            }
+        fun getIntent(context: Context): Intent {
+            return Intent(context, CreateDirectRoomActivity::class.java)
         }
     }
 }