+++ vector/src/main/java/im/vector/app/features/navigation/DefaultNavigator.kt
@@ -203,8 +203,8 @@ class DefaultNavigator @Inject constructor(
         context.startActivity(intent)
     }
 
-    override fun openCreateDirectRoom(context: Context, byQrCode: Boolean) {
-        val intent = CreateDirectRoomActivity.getIntent(context, byQrCode)
+    override fun openCreateDirectRoom(context: Context) {
+        val intent = CreateDirectRoomActivity.getIntent(context)
         context.startActivity(intent)
     }