+++ vector/src/main/java/im/vector/app/features/userdirectory/UserListViewEvents.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 New Vector Ltd
+ * Copyright (c) 2020 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.home
+package im.vector.app.features.userdirectory
 
-import im.vector.app.core.platform.VectorSharedActionViewModel
-import org.matrix.android.sdk.api.session.Session
-import javax.inject.Inject
+import im.vector.app.core.platform.VectorViewEvents
 
-class HomeSharedActionViewModel @Inject constructor(val session: Session) : VectorSharedActionViewModel<HomeActivitySharedAction>()
+/**
+ * Transient events for invite users to room screen
+ */
+sealed class UserListViewEvents : VectorViewEvents {
+    data class OpenShareMatrixToLing(val link: String) : UserListViewEvents()
+}