+++ CHANGES.md
@@ -3,6 +3,8 @@ Changes in Element 1.0.11 (2020-XX-XX)
 
 Features ✨:
  - Create DMs with users by scanning their QR code (#2025)
+ - Add Invite friends quick invite actions (#2348)
+ - Add friend by scanning QR code, show your code to friends (#2025)
 
 Improvements 🙌:
  - New room creation tile with quick action (#2346)
@@ -12,6 +14,7 @@ Improvements 🙌:
  - Handle events of type "m.room.server_acl" (#890)
  - Room creation form: add advanced section to disable federation (#1314)
  - Move "Enable Encryption" from room setting screen to room profile screen (#2394)
+ - Improve Invite user screen (seamless search for matrix ID)
 
 Bugfix 🐛:
  - Fix crash on AttachmentViewer (#2365)